+++ app/graphql/types/activity_query_type.rb
@@ -3,6 +3,7 @@
 module Types
   class ActivityQueryType < Types::BaseInputObject
     argument :user_ids, [String], required: false
+    argument :trackable_id, String, required: false
     argument :actions, [String], required: false
     argument :dates, [String], required: false
     argument :query, String, required: false