+++ db/migrate/20220421174202_add_strategic_partner_to_admin_toolkit_penetrations.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddStrategicPartnerToAdminToolkitPenetrations < ActiveRecord::Migration[6.1]
+  def change
+    safety_assured { add_column :admin_toolkit_penetrations, :strategic_partner, :string }
+  end
+end