+++ CHANGELOG.md
@@ -4,6 +4,29 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [2.2.0] - 2022-05-13
+
+### Feature
+- Trackable id support added to csv activity exports.
+
+## [2.1.0] - 2022-04-26
+
+### Features
+- Added new field strategic partners in admintoolkit penetrations.
+
+### Fixes
+- Added move in date in project level.
+
+## [2.0.1] - 2022-04-26
+
+### Fixes
+- Third_party logic update.
+
+## [2.0.0] - 2022-04-26
+
+### Features
+- Trackable id filter added to activities resolver.
+
 ## [1.9.3] - 2022-05-13
 
 ### Fixes